Moving selects to Controllers
